+++ src/main/java/org/dreambot/quests/xmarksthespot/TalkToVeosLumbridgeLeaf.java
@@ -14,16 +14,18 @@ import org.dreambot.utilities.QuestHelper;
 import org.dreambot.utilities.QuestVarBits;
 import org.dreambot.utilities.Timing;
 
-public class TalkToVeosSarimLeaf extends Leaf {
+public class TalkToVeosLumbridgeLeaf extends Leaf {
 
-    private final Area VEOS_PORT_SARIM_AREA = new Area(3050, 3249, 3055, 3245);
+    private final Area VEOS_AREA = new Area(3226, 3242, 3233, 3239);
     private final String VEOS_NAME = "Veos";
+    private final String[] DIALOGUE_OPTIONS = {"I'm looking for a quest.", "Sounds good, what should I do?", "Okay, thanks Veos.", "Yes.", "Can I help?"};
     @Override
     public boolean isValid() {
-        return PlayerSettings.getBitValue(QuestVarBits.QUEST_X_MARKS_THE_SPOT.getId()) == 6 && Inventory.contains("Ancient casket") ||
-                PlayerSettings.getBitValue(QuestVarBits.QUEST_X_MARKS_THE_SPOT.getId()) == 7;
+        return Inventory.contains("Spade") && PlayerSettings.getBitValue(QuestVarBits.QUEST_X_MARKS_THE_SPOT.getId()) == 0
+                || PlayerSettings.getBitValue(QuestVarBits.QUEST_X_MARKS_THE_SPOT.getId()) == 1
+                || PlayerSettings.getBitValue(QuestVarBits.QUEST_X_MARKS_THE_SPOT.getId()) == 2 && !Inventory.contains("Treasure scroll");
     }
 
     @Override
-    public int onLoop() { return QuestHelper.goAndTalkToNpc(VEOS_PORT_SARIM_AREA, VEOS_NAME, null); }
+    public int onLoop() { return QuestHelper.goAndTalkToNpc(VEOS_AREA, VEOS_NAME, DIALOGUE_OPTIONS); }
 }
